requirements.txt
# Core ML/DL Libraries
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Graph Libraries
networkx>=3.1
torch-geometric>=2.3.0

# Data Processing
scipy>=1.10.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Notebook and Development
jupyter>=1.0.0
ipykernel>=6.23.0

# Utilities
pyyaml>=6.0
joblib>=1.3.0

# Optional: For dataset downloading
requests>=2.31.0
